Fix involuntary rerenders in item list layout (#4720)
* Fix involuntary re-renders in item list layout



* Enable skipped integration tests for being fixed



* Remove comment


* Improve code style


* flatten file heirarchy


* Fix item list layout header rendering bug
