+++ src/renderer/components/item-object-list/index.tsx
@@ -3,4 +3,4 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 
-export * from "./item-list-layout";
+export * from "./list-layout";
